pyproject.toml
[project]
name = "llmtraining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate==0.33.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.1",
    "aiosignal==1.4.0",
    "altair==5.5.0",
    "antlr4-python3-runtime==4.9.3",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "datasets==3.0.1",
    "dill==0.3.8",
    "diskcache==5.6.3",
    "faiss-cpu==1.8.0.post1",
    "filelock==3.19.1",
    "frozenlist==1.8.0",
    "fsspec==2024.6.1",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "hf-transfer==0.1.9",
    "hf-xet==1.1.10",
    "huggingface-hub==0.35.3",
    "hydra-core==1.3.2",
    "idna==3.11",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "llama-cpp-python==0.3.16",
    "markdown-it-py==4.0.0",
    "markupsafe==2.1.5",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "multiprocess==0.70.16",
    "narwhals==2.9.0",
    "networkx==3.5",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==10.4.0",
    "platformdirs==4.5.0",
    "plotly==6.3.1",
    "pluggy==1.6.0",
    "propcache==0.4.1",
    "protobuf==5.29.5",
    "psutil==7.1.1",
    "pyarrow==21.0.0",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pytest==8.4.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "rank-bm25==0.2.2",
    "referencing==0.37.0",
    "regex==2025.10.23",
    "requests==2.32.5",
    "rich==13.9.4",
    "rpds-py==0.27.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "sentence-transformers==3.0.1",
    "setuptools==70.2.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "streamlit==1.38.0",
    "sympy==1.14.0",
    "tenacity==8.5.0",
    "threadpoolctl==3.6.0",
    "tokenizers==0.19.1",
    "toml==0.10.2",
    "torch==2.4.0",
    "torchaudio==2.4.0",
    "torchvision==0.19.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "transformers==4.44.2",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
]
